Rename envoy module to grammar for clarity
